% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{Bludger}
\alias{Bludger}
\title{Run Double Binomial GLM differential analyses}
\usage{
Bludger(se, group, contrast = c(1, 2))
}
\arguments{
\item{se}{RangedSummarizedExperiment object from `Quaffle` output}

\item{group}{Can be name of variable from `colData(se)` that contain groupings
of samples. Can also be a vector (length equal to number of samples) that
describes the sample groupings}

\item{contrast}{numeric vector of length 2 specifying which levels of
the "groups" factor should be compared.}
}
\value{
a dataframe describing the output of differential AFL usage with
the folloinwing column names:
 \itemize{
 \item{gene_id}{ID of parent gene}
 \item{gene_name}{Name of parent gene}
 \item{effectivecoord}{Coordinate of AFL event}
 \item{type}{Type of event; Alternative First (AF) or Last (AL)}
 \item{MLE_*}{Maximum likelihood estimate of event usage in each sample}
 \item{deltaMLE}{Change in event usage between comparisons}
 \item{pval}{Unadjusted p-value of differential analysis}
 \item{adj_pval}{Adjusted p-value of differential analysis}
 \item{MeanTotCt_*}{Mean total read count in each sample}
}
}
\description{
`Bludger` compares the usage of each AFL event between two groups,
using a double binomial GLM model.
}
\examples{
#' library(quaffle)
gtf <- system.file("extdata/wtap.gtf", package = "quaffle")
bams <- system.file("extdata/bams", package = "quaffle")

se <- Quaffle(bams, gtf)

Bludger(se, rep(c("A","B"), each = 3))

}
