% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4methods.R
\docType{methods}
\name{build}
\alias{build}
\alias{build,QuaffleObject-method}
\title{Build alternative first and last exon GRanges}
\usage{
build(x, ...)

\S4method{build}{QuaffleObject}(x, gtf)
}
\arguments{
\item{x}{QuaffleObject}

\item{...}{Path to input GTF}
}
\value{

}
\description{
Build alternative first and last exon GRanges
}
