% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4methods.R
\docType{methods}
\name{count}
\alias{count}
\alias{count,QuaffleObject-method}
\title{Quantify AFL reads}
\usage{
count(x, ...)

\S4method{count}{QuaffleObject}(x, dir)
}
\arguments{
\item{x}{QuaffleObject}

\item{...}{Directory containing BAM files}
}
\value{

}
\description{
Quantify AFL reads
}
