% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{createQuaffle}
\alias{createQuaffle}
\title{Create Quaffle Object}
\usage{
createQuaffle(gtf, colData, bamdir, min_read = 5)
}
\arguments{
\item{gtf}{Path to GTF annotation}

\item{colData}{DataFrame containing sample information}

\item{bamdir}{Path to directory containing BAM files. It is preferable to have
bam indices (.bam.bai) in the same directory.}

\item{min_read}{Minimum number of aligned read on AFL exons}
}
\value{
Quaffle object
}
\description{
Create Quaffle Object
}
