% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4methods.R
\docType{methods}
\name{diff}
\alias{diff}
\alias{diff,QuaffleObject-method}
\title{Perform differential AFL analysis}
\usage{
diff(x, ...)

\S4method{diff}{QuaffleObject}(x, contrast, min_samples = 2, methods = "fdr")
}
\arguments{
\item{x}{QuaffleObject}

\item{...}{Contrast vector}
}
\value{

}
\description{
Perform differential AFL analysis
}
