% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildquantdiff.R
\name{diffAFL}
\alias{diffAFL}
\title{Differential AFL analysis}
\usage{
diffAFL(psi, a, b)
}
\arguments{
\item{psi}{Dataframe with AFL PSI values from quantAFL output}

\item{a}{Character list containing sample names from group "a". Sample names
have to be identical to the name of the BAM files (without .BAM extension)}

\item{b}{Character list containing sample names from group "b". Sample names
have to be identical to the name of the BAM files (without .BAM extension)}
}
\value{
Data-frame containing differential AFL analysis
}
\description{
Differential AFL analysis
}
