% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corefuncs.R
\name{.buildAFL}
\alias{.buildAFL}
\title{Build list of alternate first and last exons}
\usage{
.buildAFL(x)
}
\arguments{
\item{x}{Path to GTF annotation}
}
\value{
GenomicRanges object with coordinates of alternative first and last exons
extracted from the input annotation
}
\description{
Build list of alternate first and last exons
}
