% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corefuncs.R
\name{.countAFL}
\alias{.countAFL}
\title{Quantify reads mapping to alternate first and last exons}
\usage{
.countAFL(x, colData, dir)
}
\arguments{
\item{x}{Quaffle object}

\item{dir}{Path to directory containing BAM files. It is preferable to have
bam indices (.bam.bai) in the same directory.}
}
\value{
Data-frame containing read count and PSI for each AFL
}
\description{
Quantify reads mapping to alternate first and last exons
}
