% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential.R
\name{.quantAFL}
\alias{.quantAFL}
\title{Quantify PSI metrics}
\usage{
.quantAFL(db, out)
}
\arguments{
\item{db}{rowRanges slot from Quaffle object}

\item{out}{counts slot from Quaffle object}
}
\value{
Matrix containing PSI values
}
\description{
Quantify PSI metrics
}
