% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsequence.R
\name{getAFLseq}
\alias{getAFLseq}
\title{Retrieve feature sequences}
\usage{
getAFLseq(
  diff,
  fasta,
  dir = NULL,
  type = c("start", "end", "exon"),
  exontype = c("all", "AF", "AL"),
  upstream = 50,
  downstream = 50,
  p.value = 0.05,
  deltapsi = 0.1,
  use.undajusted = F
)
}
\arguments{
\item{diff}{DataFrame containing differential splicing analysis, output from diff() function}

\item{fasta}{XString object containing genome sequence}

\item{dir}{Output directory for FASTA files (Optional)}

\item{type}{Type of sequence to output. Can be "start",
"end" or "exon" which will output the 5' border,
3' end or the entire exon sequence respectively.}

\item{exontype}{Type of exon to output. Can be "all" (default),
"AF" or "AL" exons.}

\item{upstream}{Amount of padding to extract for upstream sequence (Default: 50bp)}

\item{downstream}{Amount of padding to extract for downstream sequence (Default: 50bp)}

\item{p.value}{Threshold to select for statistically significant events. By default,
it will use a value of 0.05 on adjusted p-values (adj_p_val). See
`use.unadjusted` parameter to select on unadjusted p-values.}

\item{deltapsi}{Threshold to select for statistically significant events. By default,
will select events with 10% change in either samples (0.1)}

\item{use.undajusted}{Boolean value on whether to use unadjusted p-values}
}
\value{

}
\description{
Retrieve feature sequences
}
