% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4methods.R
\name{psi,QuaffleObject-method}
\alias{psi,QuaffleObject-method}
\title{Display PSI values}
\usage{
\S4method{psi}{QuaffleObject}(x)
}
\arguments{
\item{QuaffleObject}{}
}
\value{

}
\description{
Display PSI values
}
