% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{Quaffle}
\alias{Quaffle}
\title{Run Quaffle}
\usage{
Quaffle(bamdir, gtf, colData = NULL)
}
\arguments{
\item{bamdir}{Directory containing BAM files and indices}

\item{gtf}{Path to reference GTF file}

\item{colData}{Optional: A dataframe containing names of samples as rownames
and other sampel information}
}
\value{
RangedSummarizedExperiment object, with 2 assays; inc and total normalized
read counts. Number of columns in object equals to number of samples/BAM files
and number of rows equals to number of identified alternative First and Last
exons.
}
\description{
`Quaffle` will identify all alternative first and last (AFL) exons
and quantify the length-normalized read counts for each sample. This function
require a reference GTF file from which an AFL database is built upon, and
a directory containing aligned read files (BAMs). `Quaffle` outputs a
RangedSummarizedExperiment object that stores AFL coordinates, normalized
read counts and sample metadata
}
\examples{
library(quaffle)
gtf <- system.file("extdata/wtap.gtf", package = "quaffle")
bams <- system.file("extdata/bams", package = "quaffle")

se <- Quaffle(bams, gtf)
}
