% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4methods.R
\docType{class}
\name{QuaffleObject-class}
\alias{QuaffleObject-class}
\title{Quaffle Object}
\value{
Quaffle object
}
\description{
Quaffle Object
}
\section{Slots}{

\describe{
\item{\code{counts}}{Matrix containing read counts}

\item{\code{rowRanges}}{GRanges object of AFL ranges}

\item{\code{colData}}{DataFrame containing sample info}

\item{\code{psi}}{Matrix containing PSI values}

\item{\code{state}}{Matrix containing state of counts}
}}

