% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildquantdiff.R
\name{quantAFL}
\alias{quantAFL}
\title{Quantify reads mapping to alternate first and last exons}
\usage{
quantAFL(db, dir, min_read = 5)
}
\arguments{
\item{db}{Database of AFL from buildAFL output}

\item{dir}{Path to directory containing BAM files. It is preferable to have
bam indices (.bam.bai) in the same directory.}

\item{min_read}{Minimum number of aligned read on AFL exons}
}
\value{
Data-frame containing read count and PSI for each AFL
}
\description{
Quantify reads mapping to alternate first and last exons
}
