% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{Snitch}
\alias{Snitch}
\title{Get sequence}
\usage{
Snitch(
  se,
  fasta,
  subset = NULL,
  type = c("start", "end", "exon"),
  exontype = c("all", "AF", "AL"),
  upstream = 50,
  downstream = 50,
  outdir = NULL,
  filename = "Snitch.fasta"
)
}
\arguments{
\item{se}{RangedSummarizedExperiment object from `Quaffle` output}

\item{fasta}{BSGenome object or DNAStringList object containing genome sequence}

\item{subset}{A character vector containing IDs to subset. IDs can be name of
genes, ID of genes or AFL coordinates}

\item{type}{Type of sequence to return. Can be the start of exon ("start"),
end of exon ("end") or entire exon ("exon")}

\item{exontype}{Type of event to return, Can be "all", "AF" or "AL"}

\item{upstream}{Length of upstream padding sequence}

\item{downstream}{Length of downstream padding sequence}

\item{outdir}{Path to output directory (Default: only return DNAStringSet object)}

\item{filename}{Name of FASTA file to create}
}
\value{
DNAStringSet object with DNA sequences of selected events
}
\description{
Get sequence
}
